tests: Move Javascript tests to a separate file
